restructure
- Split Transport Interface implementation.
- Improve python-can Transport Interface
- Add unit tests
